+++ iotlabcli/integration/test_integration.py
@@ -1,17 +1,37 @@
 # -*- coding:utf-8 -*-
+
+# This file is a part of IoT-LAB cli-tools
+# Copyright (C) 2015 INRIA (Contact: admin@iot-lab.info)
+# Contributor(s) : see AUTHORS file
+#
+# This software is governed by the CeCILL license under French law
+# and abiding by the rules of distribution of free software.  You can  use,
+# modify and/ or redistribute the software under the terms of the CeCILL
+# license as circulated by CEA, CNRS and INRIA at the following URL
+# http://www.cecill.info.
+#
+# As a counterpart to the access to the source code and  rights to copy,
+# modify and redistribute granted by the license, users are provided only
+# with a limited warranty  and the software's author,  the holder of the
+# economic rights,  and the successive licensors  have only  limited
+# liability.
+#
+# The fact that you are presently reading this means that you have had
+# knowledge of the CeCILL license and that you accept its terms.
+
 """ Integration tests for cli-tools """
 
 # pylint:disable=I0011,too-many-public-methods
 # pylint:disable=I0011,attribute-defined-outside-init
 # pylint:disable=I0011,invalid-name
 
-#
+
 # Test that may be added:
 #  * physical experiments
 #  * load experiment
 #  * Validates JSON outputs
 #  * run commands without exp_id in error cases no exp or many exps running
-#
+
 
 import argparse
 import os